+++ src/components/ids-data-grid/demos/auto-fit-bottom.ts
@@ -2,7 +2,7 @@ import type IdsDataGrid from '../ids-data-grid';
 import '../ids-data-grid';
 import type { IdsDataGridColumn } from '../ids-data-grid-column';
 import productsJSON from '../../../assets/data/products.json';
-
+import '../../ids-layout-flex/ids-layout-flex';
 import css from '../../../assets/css/ids-data-grid/auto-fit.css';
 
 const cssLink = `<link href="${css}" rel="stylesheet">`;